+++ src/index.ts
@@ -1,3 +1,4 @@
 import * as apollo from "./apollo"
 export { apollo }
 export { default as query } from "./query"
+export { default as mutation } from "./mutation"