+++ src/types.ts
@@ -7,3 +7,11 @@ export interface QueryAttributes<Data = {}, Variables = {}> {
   loading: boolean
   refetch: () => void
 }
+
+export interface MutationAttributes<Data = {}, Variables = {}> {
+  data: Data | null | undefined
+  errors: GraphQLError[] | null | undefined
+  loading: boolean
+  called: boolean
+  execute: (data: { variables?: Variables }) => void
+}