+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This change log adheres to [keepachangelog.com](http://keepachangelog.com).
 ### Added
 - Add hyperapp-apollo.js and hyperapp-apollo.min.js to dist files.
 - Enable to use `<Query/>` in a loop.
+- Add `mutation()` and `<Mutation/>` component.
 
 ### Changed
 - Move hyperapp from peerDependencies to dependencies.